Broke utils into multiple file, logical consolidation of constants into different files, reworked server state and corresponding update commands.
