+++ container_service_extension/cluster.py
@@ -18,10 +18,10 @@ from container_service_extension.exceptions import CseServerError
 from container_service_extension.exceptions import DeleteNodeError
 from container_service_extension.exceptions import NodeCreationError
 from container_service_extension.exceptions import ScriptExecutionError
+from container_service_extension.install_utils import get_data_file
+from container_service_extension.install_utils import get_vsphere
 from container_service_extension.logger import SERVER_LOGGER as LOGGER
-from container_service_extension.utils import get_data_file
-from container_service_extension.utils import get_vsphere
-from container_service_extension.utils import SYSTEM_ORG_NAME
+from container_service_extension.server_constants import SYSTEM_ORG_NAME
 
 TYPE_MASTER = 'mstr'
 TYPE_NODE = 'node'
