+++ container_service_extension/shared_constants.py
@@ -1,0 +1,14 @@
+# container-service-extension
+# Copyright (c) 2017 VMware, Inc. All Rights Reserved.
+# SPDX-License-Identifier: BSD-2-Clause
+
+
+ERROR_DESCRIPTION = "description"
+ERROR_MESSAGE = "message"
+ERROR_REASON = "reason"
+ERROR_STACKTRACE = "stacktrace"
+
+SERVER_ACTION_KEY = 'action'
+SERVER_DISABLE_ACTION = 'disable'
+SERVER_ENABLE_ACTION = 'enable'
+SERVER_STOP_ACTION = 'stop'
